Upgrading math for new functionality
Unit tests now passing. (The were broken with the prior commit).

Note we've added new functionality, but we've not fully tested it. We are only demonstrating that the new functionality is consistent with the 0.0.1 functionality.  We know it works on one commodity min cost network flow, so now we can test it with true multi-commodity network flow problems.